refactor: 波形合成テストの移植
