+++ test/tts_pipeline/test_tts_engine.py
@@ -10,22 +10,12 @@ from syrupy.extensions.json import JSONSnapshotExtension
 
 from voicevox_engine.dev.core.mock import MockCoreWrapper
 from voicevox_engine.metas.Metas import StyleId
-from voicevox_engine.model import AccentPhrase, AudioQuery, Mora
+from voicevox_engine.model import AccentPhrase, Mora
 from voicevox_engine.tts_pipeline.acoustic_feature_extractor import Phoneme
 from voicevox_engine.tts_pipeline.text_analyzer import text_to_accent_phrases
 from voicevox_engine.tts_pipeline.tts_engine import (
     TTSEngine,
     apply_interrogative_upspeak,
-    apply_intonation_scale,
-    apply_output_sampling_rate,
-    apply_output_stereo,
-    apply_pitch_scale,
-    apply_prepost_silence,
-    apply_speed_scale,
-    apply_volume_scale,
-    count_frame_per_unit,
-    query_to_decoder_feature,
-    raw_wave_to_output_wave,
     split_mora,
     to_flatten_moras,
     to_flatten_phonemes,
@@ -117,32 +107,6 @@ class MockCore:
         return True
 
 
-def _gen_query(
-    accent_phrases: list[AccentPhrase] | None = None,
-    speedScale: float = 1.0,
-    pitchScale: float = 1.0,
-    intonationScale: float = 1.0,
-    prePhonemeLength: float = 0.0,
-    postPhonemeLength: float = 0.0,
-    volumeScale: float = 1.0,
-    outputSamplingRate: int = 24000,
-    outputStereo: bool = False,
-) -> AudioQuery:
-    """Generate AudioQuery with default meaningless arguments for test simplicity."""
-    accent_phrases = [] if accent_phrases is None else accent_phrases
-    return AudioQuery(
-        accent_phrases=accent_phrases,
-        speedScale=speedScale,
-        pitchScale=pitchScale,
-        intonationScale=intonationScale,
-        prePhonemeLength=prePhonemeLength,
-        postPhonemeLength=postPhonemeLength,
-        volumeScale=volumeScale,
-        outputSamplingRate=outputSamplingRate,
-        outputStereo=outputStereo,
-    )
-
-
 def _gen_mora(
     text: str,
     consonant: str | None,
@@ -178,274 +142,6 @@ def test_to_flatten_phonemes():
     phonemes = list(map(lambda p: p.phoneme, to_flatten_phonemes(moras)))
 
     assert true_phonemes == phonemes
-
-
-def test_apply_prepost_silence():
-    """Test `apply_prepost_silence`."""
-    # Inputs
-    query = _gen_query(prePhonemeLength=2 * 0.01067, postPhonemeLength=6 * 0.01067)
-    moras = [
-        _gen_mora("ヒ", "h", 2 * 0.01067, "i", 4 * 0.01067, 100.0),
-    ]
-
-    # Expects
-    true_moras_with_silence = [
-        _gen_mora("　", None, None, "sil", 2 * 0.01067, 0.0),
-        _gen_mora("ヒ", "h", 2 * 0.01067, "i", 4 * 0.01067, 100.0),
-        _gen_mora("　", None, None, "sil", 6 * 0.01067, 0.0),
-    ]
-
-    # Outputs
-    moras_with_silence = apply_prepost_silence(moras, query)
-
-    assert moras_with_silence == true_moras_with_silence
-
-
-def test_apply_speed_scale():
-    """Test `apply_speed_scale`."""
-    # Inputs
-    query = _gen_query(speedScale=2.0)
-    input_moras = [
-        _gen_mora("コ", "k", 2 * 0.01067, "o", 4 * 0.01067, 50.0),
-        _gen_mora("ン", None, None, "N", 4 * 0.01067, 50.0),
-        _gen_mora("、", None, None, "pau", 2 * 0.01067, 0.0),
-        _gen_mora("ヒ", "h", 2 * 0.01067, "i", 4 * 0.01067, 125.0),
-        _gen_mora("ホ", "h", 4 * 0.01067, "O", 2 * 0.01067, 0.0),
-    ]
-
-    # Expects - x2 fast
-    true_moras = [
-        _gen_mora("コ", "k", 1 * 0.01067, "o", 2 * 0.01067, 50.0),
-        _gen_mora("ン", None, None, "N", 2 * 0.01067, 50.0),
-        _gen_mora("、", None, None, "pau", 1 * 0.01067, 0.0),
-        _gen_mora("ヒ", "h", 1 * 0.01067, "i", 2 * 0.01067, 125.0),
-        _gen_mora("ホ", "h", 2 * 0.01067, "O", 1 * 0.01067, 0.0),
-    ]
-
-    # Outputs
-    moras = apply_speed_scale(input_moras, query)
-
-    assert moras == true_moras
-
-
-def test_apply_pitch_scale():
-    """Test `apply_pitch_scale`."""
-    # Inputs
-    query = _gen_query(pitchScale=2.0)
-    input_moras = [
-        _gen_mora("コ", "k", 0.0, "o", 0.0, 50.0),
-        _gen_mora("ン", None, None, "N", 0.0, 50.0),
-        _gen_mora("、", None, None, "pau", 0.0, 0.0),
-        _gen_mora("ヒ", "h", 0.0, "i", 0.0, 125.0),
-        _gen_mora("ホ", "h", 0.0, "O", 0.0, 0.0),
-    ]
-
-    # Expects - x4 value scaled
-    true_moras = [
-        _gen_mora("コ", "k", 0.0, "o", 0.0, 200.0),
-        _gen_mora("ン", None, None, "N", 0.0, 200.0),
-        _gen_mora("、", None, None, "pau", 0.0, 0.0),
-        _gen_mora("ヒ", "h", 0.0, "i", 0.0, 500.0),
-        _gen_mora("ホ", "h", 0.0, "O", 0.0, 0.0),
-    ]
-
-    # Outputs
-    moras = apply_pitch_scale(input_moras, query)
-
-    assert moras == true_moras
-
-
-def test_apply_intonation_scale():
-    """Test `apply_intonation_scale`."""
-    # Inputs
-    query = _gen_query(intonationScale=0.5)
-    input_moras = [
-        _gen_mora("コ", "k", 0.0, "o", 0.0, 200.0),
-        _gen_mora("ン", None, None, "N", 0.0, 200.0),
-        _gen_mora("、", None, None, "pau", 0.0, 0.0),
-        _gen_mora("ヒ", "h", 0.0, "i", 0.0, 500.0),
-        _gen_mora("ホ", "h", 0.0, "O", 0.0, 0.0),
-    ]
-
-    # Expects - mean=300 var x0.5 intonation scaling
-    true_moras = [
-        _gen_mora("コ", "k", 0.0, "o", 0.0, 250.0),
-        _gen_mora("ン", None, None, "N", 0.0, 250.0),
-        _gen_mora("、", None, None, "pau", 0.0, 0.0),
-        _gen_mora("ヒ", "h", 0.0, "i", 0.0, 400.0),
-        _gen_mora("ホ", "h", 0.0, "O", 0.0, 0.0),
-    ]
-
-    # Outputs
-    moras = apply_intonation_scale(input_moras, query)
-
-    assert moras == true_moras
-
-
-def test_apply_volume_scale():
-    """Test `apply_volume_scale`."""
-    # Inputs
-    query = _gen_query(volumeScale=3.0)
-    input_wave = numpy.array([0.0, 1.0, 2.0])
-
-    # Expects - x3 scale
-    true_wave = numpy.array([0.0, 3.0, 6.0])
-
-    # Outputs
-    wave = apply_volume_scale(input_wave, query)
-
-    assert numpy.allclose(wave, true_wave)
-
-
-def test_apply_output_sampling_rate():
-    """Test `apply_output_sampling_rate`."""
-    # Inputs
-    query = _gen_query(outputSamplingRate=12000)
-    input_wave = numpy.array([1.0 for _ in range(120)])
-    input_sr_wave = 24000
-
-    # Expects - half sampling rate
-    true_wave = numpy.array([1.0 for _ in range(60)])
-    assert true_wave.shape == (60,), "Prerequisites"
-
-    # Outputs
-    wave = apply_output_sampling_rate(input_wave, input_sr_wave, query)
-
-    assert wave.shape[0] == true_wave.shape[0]
-
-
-def test_apply_output_stereo():
-    """Test `apply_output_stereo`."""
-    # Inputs
-    query = _gen_query(outputStereo=True)
-    input_wave = numpy.array([1.0, 0.0, 2.0])
-
-    # Expects - Stereo :: (Time, Channel)
-    true_wave = numpy.array([[1.0, 1.0], [0.0, 0.0], [2.0, 2.0]])
-
-    # Outputs
-    wave = apply_output_stereo(input_wave, query)
-
-    assert numpy.array_equal(wave, true_wave)
-
-
-def test_count_frame_per_unit():
-    """Test `count_frame_per_unit`."""
-    # Inputs
-    moras = [
-        _gen_mora("　", None, None, "　", 2 * 0.01067, 0.0),  # 0.01067 [sec/frame]
-        _gen_mora("コ", "k", 2 * 0.01067, "o", 4 * 0.01067, 0.0),
-        _gen_mora("ン", None, None, "N", 4 * 0.01067, 0.0),
-        _gen_mora("、", None, None, "pau", 2 * 0.01067, 0.0),
-        _gen_mora("ヒ", "h", 2 * 0.01067, "i", 4 * 0.01067, 0.0),
-        _gen_mora("ホ", "h", 4 * 0.01067, "O", 2 * 0.01067, 0.0),
-        _gen_mora("　", None, None, "　", 6 * 0.01067, 0.0),
-    ]
-
-    # Expects
-    #                             Pre k  o  N pau h  i  h  O Pst
-    true_frame_per_phoneme_list = [2, 2, 4, 4, 2, 2, 4, 4, 2, 6]
-    true_frame_per_phoneme = numpy.array(true_frame_per_phoneme_list, dtype=numpy.int32)
-    #                         Pre ko  N pau hi hO Pst
-    true_frame_per_mora_list = [2, 6, 4, 2, 6, 6, 6]
-    true_frame_per_mora = numpy.array(true_frame_per_mora_list, dtype=numpy.int32)
-
-    # Outputs
-    frame_per_phoneme, frame_per_mora = count_frame_per_unit(moras)
-
-    assert numpy.array_equal(frame_per_phoneme, true_frame_per_phoneme)
-    assert numpy.array_equal(frame_per_mora, true_frame_per_mora)
-
-
-def test_query_to_decoder_feature():
-    """Test `query_to_decoder_feature`."""
-    # Inputs
-    accent_phrases = [
-        AccentPhrase(
-            moras=[
-                _gen_mora("コ", "k", 2 * 0.01067, "o", 4 * 0.01067, 50.0),
-                _gen_mora("ン", None, None, "N", 4 * 0.01067, 50.0),
-            ],
-            accent=1,
-            pause_mora=_gen_mora("、", None, None, "pau", 2 * 0.01067, 0.0),
-        ),
-        AccentPhrase(
-            moras=[
-                _gen_mora("ヒ", "h", 2 * 0.01067, "i", 4 * 0.01067, 125.0),
-                _gen_mora("ホ", "h", 4 * 0.01067, "O", 2 * 0.01067, 0.0),
-            ],
-            accent=1,
-            pause_mora=None,
-        ),
-    ]
-    query = _gen_query(
-        accent_phrases=accent_phrases,
-        speedScale=2.0,
-        pitchScale=2.0,
-        intonationScale=0.5,
-        prePhonemeLength=2 * 0.01067,
-        postPhonemeLength=6 * 0.01067,
-    )
-
-    # Expects
-    # frame_per_phoneme
-    #                        Pre k  o  N pau h  i  h  O Pst
-    true_frame_per_phoneme = [1, 1, 2, 2, 1, 1, 2, 2, 1, 3]
-    n_frame = sum(true_frame_per_phoneme)
-    # phoneme
-    #                     Pr  k   o   o  N  N pau  h   i   i   h   h  O Pt Pt Pt
-    frame_phoneme_idxs = [0, 23, 30, 30, 4, 4, 0, 19, 21, 21, 19, 19, 5, 0, 0, 0]
-    true_phoneme = numpy.zeros([n_frame, TRUE_NUM_PHONEME], dtype=numpy.float32)
-    for frame_idx, phoneme_idx in enumerate(frame_phoneme_idxs):
-        true_phoneme[frame_idx, phoneme_idx] = 1.0
-    # Pitch
-    #                   paw ko  N pau hi hO paw
-    # frame_per_vowel = [1, 3,  2, 1, 3, 3, 3]
-    #           pau   ko     ko     ko      N      N
-    true1_f0 = [0.0, 250.0, 250.0, 250.0, 250.0, 250.0]
-    #           pau   hi     hi     hi
-    true2_f0 = [0.0, 400.0, 400.0, 400.0]
-    #           hO   hO   hO   paw  paw  paw
-    true3_f0 = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-    true_f0 = numpy.array(true1_f0 + true2_f0 + true3_f0, dtype=numpy.float32)
-
-    # Outputs
-    phoneme, f0 = query_to_decoder_feature(query)
-
-    assert numpy.array_equal(phoneme, true_phoneme)
-    assert numpy.array_equal(f0, true_f0)
-
-
-def test_raw_wave_to_output_wave_with_resample():
-    """Test `raw_wave_to_output_wave` with resampling option."""
-    # Inputs
-    query = _gen_query(volumeScale=2, outputSamplingRate=48000, outputStereo=True)
-    raw_wave = numpy.random.rand(240)
-    sr_raw_wave = 24000
-
-    # Expects
-    true_wave_shape = (480, 2)
-
-    # Outputs
-    wave = raw_wave_to_output_wave(query, raw_wave, sr_raw_wave)
-
-    assert wave.shape == true_wave_shape
-
-
-def test_raw_wave_to_output_wave_without_resample():
-    """Test `raw_wave_to_output_wave`  without resampling option."""
-    # Inputs
-    query = _gen_query(volumeScale=2, outputStereo=True)
-    raw_wave = numpy.random.rand(240)
-    sr_raw_wave = 24000
-
-    # Expects
-    true_wave = numpy.array([2 * raw_wave, 2 * raw_wave]).T
-
-    # Outputs
-    wave = raw_wave_to_output_wave(query, raw_wave, sr_raw_wave)
-
-    assert numpy.allclose(wave, true_wave)
 
 
 def _gen_hello_hiho_accent_phrases() -> list[AccentPhrase]:
